+++ Tests/MapboxMapsTests/Foundation/Mocks/MockMapboxMap.swift
@@ -290,7 +290,7 @@ final class MockMapboxMap: MapboxMapProtocol {
         guard let featureset = interaction.featureset else { return AnyCancelable.empty }
         return addInteraction(InteractionImpl(featureset: .init(core: featureset), filter: nil, type: type, onBegin: { feature, context in
             let queriedFeature = QueriedFeature(
-                __feature: MapboxCommon.Feature(feature.geoJsonFeature),
+                __feature: MapboxCommon.Feature(feature.originalFeature),
                 source: "",
                 sourceLayer: nil,
                 state: [String: Any](),
@@ -334,7 +334,7 @@ final class MockMapboxMap: MapboxMapProtocol {
             FeaturesetFeature(
                 id: feature.identifier?.string.map { FeaturesetFeatureId(id: $0) },
                 featureset: featureset,
-                geoJsonFeature: feature,
+                originalFeature: feature,
                 state: JSONObject())
         } else {
             nil